Home page with product and name of the product is cliquable

+++ DB/DB_Projet_ASP/Tables/Produit.sql
@@ -3,7 +3,7 @@
 	[Id_Prod] INT IDENTITY NOT NULL PRIMARY KEY, 
     [nom] NVARCHAR(64) NOT NULL, 
     [description] NVARCHAR(MAX) NULL, 
-    [prix] DECIMAL(10, 2) NOT NULL, 
+    [prix] FLOAT NOT NULL, 
     [cri_eco] NVARCHAR(64) NULL, 
     [Id_Cat] INT NOT NULL, 
     CONSTRAINT [FK_Produit_Categories] FOREIGN KEY ([Id_Cat]) REFERENCES [Categories]([Id_Cat]),
